move json write to script file, move api fetch and transform functions to helper file
