+++ src/utils/index.js
@@ -2,3 +2,4 @@ export * from "./casing";
 export * from "./download-csv";
 export * from "./download-json";
 export * from "./log-logo";
+export * from "./studies-data-helpers";
